♻️ Refactor: callchat 엔티티 파일 이동
