+++ src/main/java/callprotector/spring/domain/chat/service/CallChatSessionServiceImpl.java
@@ -2,7 +2,7 @@ package callprotector.spring.domain.chat.service;
 
 import callprotector.spring.domain.callsession.entity.CallSession;
 import callprotector.spring.domain.callchat.dto.response.CallChatSessionResponseDTO;
-import callprotector.spring.domain.chat.entity.CallChatSession;
+import callprotector.spring.domain.callchat.entity.CallChatSession;
 import callprotector.spring.domain.chat.repository.CallChatSessionRepository;
 import callprotector.spring.domain.user.entity.User;
 import org.springframework.transaction.annotation.Transactional;
